split sp and idp
