+++ mslib/auth_client_sp/README.md
