+++ mslib/auth_client_sp/app/templates/base.html
