+++ mslib/auth_client_sp/app/templates/index.html
